+++ src/binary_kind.mli
@@ -1,0 +1,10 @@
+(** Linking modes for binaries *)
+
+type t =
+  | Exe
+  | Object
+  | Shared_object
+
+val t : t Sexp.Of_sexp.t
+
+val all : t list