+++ src/exe.ml
@@ -45,6 +45,54 @@ module Linkage = struct
     ; ext
     ; flags
     }
+
+  let  o_flags = ["-output-complete-obj"]
+  let so_flags = ["-output-complete-obj"; "-runtime-variant"; "_pic"]
+
+  let of_user_config (ctx : Context.t) (m : Jbuild.Executables.Link_mode.t) =
+    let wanted_mode : Mode.t =
+      match m.mode with
+      | Byte   -> Byte
+      | Native -> Native
+      | Best   -> Native
+    in
+    let real_mode : Mode.t =
+      match m.mode with
+      | Byte   -> Byte
+      | Native -> Native
+      | Best   -> if Option.is_some ctx.ocamlopt then Native else Byte
+    in
+    let ext =
+      match wanted_mode, m.kind with
+      | Byte   , Exe           -> ".bc"
+      | Native , Exe           -> ".exe"
+      | Byte   , Object        -> ".bc"  ^ ctx.ext_obj
+      | Native , Object        -> ".exe" ^ ctx.ext_obj
+      | Byte   , Shared_object -> ".bc"  ^ ctx.ext_dll
+      | Native , Shared_object ->          ctx.ext_dll
+    in
+    let flags =
+      match m.kind with
+      | Exe ->
+        if wanted_mode = Native && real_mode = Byte then
+          ["-custom"]
+        else
+          []
+      | Object -> o_flags
+      | Shared_object ->
+        if real_mode = Native then
+          (* The compiler doesn't pass these flags in native mode. This
+             looks like a bug in the compiler. *)
+          List.concat_map ctx.native_c_libraries ~f:(fun flag ->
+            ["-cclib"; flag])
+          @ so_flags
+        else
+          so_flags
+    in
+    { ext
+    ; mode = real_mode
+    ; flags
+    }
 end
 
 let link_exe
@@ -101,7 +149,7 @@ let link_exe
            Lib.L.link_flags libs ~mode ~stdlib_dir:ctx.stdlib_dir)
        ; Dyn (fun (_, cm_files, _, _) -> Deps cm_files)
        ]);
-  if mode = Mode.Byte then
+  if linkage.ext = ".bc" then
     let rules = Js_of_ocaml_rules.build_exe sctx ~dir ~js_of_ocaml ~src:exe in
     let libs_and_cm_and_flags =
       (requires &&& (modules_and_cm_files >>^ snd))