+++ CHANGES.md
@@ -35,6 +35,11 @@ next
 - Make sure modules in the current directory always have precedence
   over included directories (#597)
 
+- Add support for building executables as object or shared object
+  files (#23)
+
+- Add a `best` mode which is native with fallback to byte-code when
+  native compilation is not available (#23)
 
 1.0+beta18 (25/02/2018)
 -----------------------